# doc/restapi/restapi.swagger2.json
{
  "swagger": "2.0",
  "info": {
    "version": "v1",
    "title": "API V1"
  },
  "basePath": "/",
  "paths": {
...
    "/api/Leaderboards/global/list": {
      "get": {
        "tags": [ "Leaderboards" ],
        "summary": "Find a list of leaderboards that match <param name=\"gameId\" />.\r\n            If global is provided instead of a gameId, get all global leaderboards, ie. leaderboards that are not associated with a specific game.\r\n            \r\n            Example Usage: GET api/leaderboards/game/1/list",
        "operationId": "ApiLeaderboardsGlobalListGet",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "description": "Game ID",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": { "204": { "description": "No Content" } },
        "deprecated": false
      }
    },
    "/api/Leaderboards/game/{gameId}/list": {
      "get": {
        "tags": [ "Leaderboards" ],
        "summary": "Find a list of leaderboards that match <param name=\"gameId\" />.\r\n            If global is provided instead of a gameId, get all global leaderboards, ie. leaderboards that are not associated with a specific game.\r\n            \r\n            Example Usage: GET api/leaderboards/game/1/list",
        "operationId": "ApiLeaderboardsGameByGameIdListGet",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "description": "Game ID",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": { "204": { "description": "No Content" } },
        "deprecated": false
      }
    },
    "/api/Leaderboards/{token}/{gameId}": {
      "get": {
        "tags": [ "Leaderboards" ],
        "summary": "Find a single leaderboard matching the token and gameId.\r\n            \r\n            Example Usage: GET api/leaderboards/LEADERBOARD_TOKEN/1",
        "operationId": "ApiLeaderboardsByTokenByGameIdGet",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "description": "Token",
            "required": true,
            "type": "string"
          },
          {
            "name": "gameId",
            "in": "path",
            "description": "",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": { "204": { "description": "No Content" } },
        "deprecated": false
      },
      "delete": {
        "tags": [ "Leaderboards" ],
        "summary": "Delete Leaderboard with the <param name=\"token\" /> and <param name=\"gameId\" /> provided.\r\n            \r\n            Example Usage: DELETE api/leaderboards/LEADERBOARD_TOKEN/1",
        "operationId": "ApiLeaderboardsByTokenByGameIdDelete",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "description": "Token of Leaderboard",
            "required": true,
            "type": "string"
          },
          {
            "name": "gameId",
            "in": "path",
            "description": "ID of the Game the Leaderboard is for",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": { "204": { "description": "No Content" } },
        "deprecated": false
      }
    },
    "/api/Leaderboards/create": {
      "post": {
        "tags": [ "Leaderboards" ],
        "summary": "Create a new Leaderboard.\r\n            Requires {PlayGen.SUGAR.Contracts.LeaderboardRequest.Name} and {PlayGen.SUGAR.Contracts.LeaderboardRequest.Token} to be unique to that {PlayGen.SUGAR.Contracts.LeaderboardRequest.GameId}.\r\n            \r\n            Example Usage: POST api/leaderboards/create",
        "operationId": "ApiLeaderboardsCreatePost",
        "consumes": [ "application/json", "text/json", "application/json-patch+json" ],
        "produces": [],
        "parameters": [
          {
            "name": "newLeaderboard",
            "in": "body",
            "description": "{PlayGen.SUGAR.Contracts.LeaderboardRequest} object that holds the details of the new Leaderboard.",
            "required": false,
            "schema": { "$ref": "#/definitions/LeaderboardRequest" }
          }
        ],
        "responses": { "204": { "description": "No Content" } },
        "deprecated": false
      }
    },
    "/api/Leaderboards/standings": {
      "post": {
        "tags": [ "Leaderboards" ],
        "summary": "Get the standings for a Leaderboard using a {PlayGen.SUGAR.Contracts.LeaderboardStandingsRequest}.\r\n            \r\n            Example Usage: POST api/leaderboards/standings",
        "operationId": "ApiLeaderboardsStandingsPost",
        "consumes": [ "application/json", "text/json", "application/json-patch+json" ],
        "produces": [],
        "parameters": [
          {
            "name": "leaderboardDetails",
            "in": "body",
            "description": "{PlayGen.SUGAR.Contracts.LeaderboardStandingsRequest} object that holds the details that are wanted from the Leaderboard.",
            "required": false,
            "schema": { "$ref": "#/definitions/LeaderboardStandingsRequest" }
          }
        ],
        "responses": { "204": { "description": "No Content" } },
        "deprecated": false
      }
    },
    "/api/Leaderboards/update": {
      "put": {
        "tags": [ "Leaderboards" ],
        "summary": "Update an existing Leaderboard.\r\n            \r\n            Example Usage: PUT api/leaderboards/update",
        "operationId": "ApiLeaderboardsUpdatePut",
        "consumes": [ "application/json", "text/json", "application/json-patch+json" ],
        "produces": [],
        "parameters": [
          {
            "name": "leaderboard",
            "in": "body",
            "description": "{PlayGen.SUGAR.Contracts.LeaderboardRequest} object that holds the details of the Leaderboard.",
            "required": false,
            "schema": { "$ref": "#/definitions/LeaderboardRequest" }
          }
        ],
        "responses": { "204": { "description": "No Content" } },
        "deprecated": false
      }
    },
    "/api/Leaderboards/{token}/global": {
      "delete": {
        "tags": [ "Leaderboards" ],
        "summary": "Delete Leaderboard with the <param name=\"token\" /> and <param name=\"gameId\" /> provided.\r\n            \r\n            Example Usage: DELETE api/leaderboards/LEADERBOARD_TOKEN/1",
        "operationId": "ApiLeaderboardsByTokenGlobalDelete",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "description": "Token of Leaderboard",
            "required": true,
            "type": "string"
          },
          {
            "name": "gameId",
            "in": "path",
            "description": "ID of the Game the Leaderboard is for",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": { "204": { "description": "No Content" } },
        "deprecated": false
      }
    },
...
  },
  "definitions": {
...
    "LeaderboardRequest": {
      "description": "Encapsulates leaderboard details.",
      "required": [ "token", "name", "actorType", "gameDataType", "criteriaScope", "leaderboardType" ],
      "type": "object",
      "properties": {
        "token": {
          "description": "A unique identifier used in development to reference the leaderboard.",
          "maxLength": 64,
          "minLength": 0,
          "type": "string"
        },
        "gameId": {
          "format": "int32",
          "description": "The Id of the Game which this leaderboards belongs to. Can be left null to relate this leaderboard to the system rather than a game.\r\n            This Id is also used when gathering GameData using the Key provided.",
          "type": "integer"
        },
        "name": {
          "description": "The display name of the leaderboard.",
          "maxLength": 64,
          "minLength": 0,
          "type": "string"
        },
        "key": {
          "description": "The GameData key which is checked against in order to create the leaderboard standings.",
          "maxLength": 64,
          "minLength": 0,
          "type": "string"
        },
        "actorType": {
          "description": "The type of actor which this leaderboard is intended for.",
          "enum": [ "Undefined", "User", "Group" ],
          "type": "string"
        },
        "gameDataType": {
          "description": "The GameDataType of the GameData being checked against.",
          "enum": [ "String", "Long", "Float", "Boolean" ],
          "type": "string"
        },
        "criteriaScope": {
          "description": "Whether the criteria will be checked against the actor or relatedactors (i.e. group members, user friends).",
          "enum": [ "Actor", "RelatedActors" ],
          "type": "string"
        },
        "leaderboardType": {
          "description": "The method which data will be collected and sorted for this leaderboard.",
          "enum": [ "Highest", "Lowest", "Cumulative", "Count", "Earliest", "Latest" ],
          "type": "string"
        }
      }
    },
    "LeaderboardStandingsRequest": {
      "description": "Encapsulates leaderboard current standings request.",
      "required": [ "leaderboardToken", "gameId", "leaderboardFilterType", "limit", "offset" ],
      "type": "object",
      "properties": {
        "leaderboardToken": {
          "description": "The Token of the Leaderboard which the standings are being gathered for.",
          "type": "string"
        },
        "gameId": {
          "format": "int32",
          "description": "The Id of the Game the leaderboard. Can be left null for system-wide leaderboards.",
          "type": "integer"
        },
        "actorId": {
          "format": "int32",
          "description": "The Id of an Actor. Required for getting standings for Near, Friends (user only) and Group Members (group only)",
          "type": "integer"
        },
        "leaderboardFilterType": {
          "description": "The filter for what standings will be returned.",
          "enum": [ "Top", "Near", "Friends", "GroupMembers" ],
          "type": "string"
        },
        "limit": {
          "format": "int32",
          "description": "The maximum number of results which will be returned.",
          "type": "integer"
        },
        "offset": {
          "format": "int32",
          "description": "The set of results which will be returned based on the limit.\r\n            0 returns the first set of results for Top, Friends and Group Member and the nearest range for Near.",
          "type": "integer"
        },
        "dateStart": {
          "format": "date-time",
          "description": "The earliest point which data should be collected from.\r\n            Can be left null to set no earliest point.",
          "type": "string"
        },
        "dateEnd": {
          "format": "date-time",
          "description": "The latest point which data should be collected from.\r\n            Can be left null to gather ll data up to the present time.",
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {}
}